requirements.txt
streamlit==1.10.0
llama-cpp-python==0.1.0
groq==0.1.0
plotly==5.5.0
pyyaml==5.4.1
bcrypt==3.2.0